// packages/server/src/Interface.ts
export type ExecutionState = 'INPROGRESS' | 'FINISHED' | 'ERROR' | 'TERMINATED' | 'TIMEOUT' | 'STOPPED'
// ...
export interface IExecution {
    id: string
    executionData: string
    state: ExecutionState
    agentflowId: string
    sessionId: string
    isPublic?: boolean
    action?: string
    createdDate: Date
    updatedDate: Date
    stoppedDate: Date
    workspaceId: string
}